requirements.txt
Flask==2.2.2
imageio==2.28.0
numpy==1.23.5
pandas==1.5.2
python-dotenv==1.0.0
Requests==2.29.0
scikit_learn==1.2.2
